requirements.txt
langchain
langchain_community
wikipedia
pydantic>=1.10,<2.0
python-dotenv
requests  # Ensure HTTP requests work if needed
